# contratos/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto_cusco API",
    "description": "API REST del backend de auto_cusco: ingesta de sabanas, seleccion y segmentacion de cartera, generacion de cargas (digitales y VoIP) y reportes de cobranza.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cartera": {
      "get": {
        "tags": [
          "cartera"
        ],
        "summary": "Consultar Cartera",
        "description": "Productos de la version vigente de la fecha (RF-04, RF-05, RF-08).",
        "operationId": "consultar_cartera_cartera_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Corte"
            }
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Filtro"
            }
          },
          {
            "name": "orden",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "campo, o -campo para descendente",
              "title": "Orden"
            },
            "description": "campo, o -campo para descendente"
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cartera/metricas": {
      "get": {
        "tags": [
          "cartera"
        ],
        "summary": "Metricas Cartera",
        "description": "Capital, cuentas, segmentos y cuotas de la seleccion (RF-26), mas RF-27.",
        "operationId": "metricas_cartera_cartera_metricas_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Corte"
            }
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Filtro"
            }
          },
          {
            "name": "indicador",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Indicador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricasRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cartera/segmentacion": {
      "get": {
        "tags": [
          "cartera"
        ],
        "summary": "Segmentar Cartera",
        "description": "Cuentas y capital por cada valor de un atributo (RF-06).",
        "operationId": "segmentar_cartera_cartera_segmentacion_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Corte"
            }
          },
          {
            "name": "campo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campo"
            }
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Filtro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentacionRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cartera/resumen": {
      "get": {
        "tags": [
          "cartera"
        ],
        "summary": "Resumir Cartera",
        "description": "Metricas del universo filtrado y de sus primeros n, lado a lado (RF-08, RF-26 a RF-28).\n\nLos primeros n son los mismos productos que devuelve `/cartera` con ese orden\ny los mismos que se escriben en el archivo de carga.",
        "operationId": "resumir_cartera_cartera_resumen_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Corte"
            }
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Filtro"
            }
          },
          {
            "name": "orden",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Que productos son los primeros: campo, o -campo",
              "title": "Orden"
            },
            "description": "Que productos son los primeros: campo, o -campo"
          },
          {
            "name": "cantidad",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "El n del top n. Sin ella no hay bloque de seleccion",
              "title": "Cantidad"
            },
            "description": "El n del top n. Sin ella no hay bloque de seleccion"
          },
          {
            "name": "indicador",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Indicador"
            }
          },
          {
            "name": "segmento",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Campo por el que segmentar los dos bloques",
              "title": "Segmento"
            },
            "description": "Campo por el que segmentar los dos bloques"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumenRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BloqueRespuesta": {
        "properties": {
          "metricas": {
            "$ref": "#/components/schemas/MetricasRespuesta"
          },
          "segmentacion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SegmentacionRespuesta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "metricas",
          "segmentacion"
        ],
        "title": "BloqueRespuesta"
      },
...
      "DetalleError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "DetalleError",
        "description": "Motivo legible por el que la API rechazo la peticion."
      },
...
      "GrupoRespuesta": {
        "properties": {
          "valor": {
            "title": "Valor"
          },
          "cuentas": {
            "type": "integer",
            "title": "Cuentas"
          },
          "capital": {
            "type": "string",
            "title": "Capital"
          }
        },
        "type": "object",
        "required": [
          "valor",
          "cuentas",
          "capital"
        ],
        "title": "GrupoRespuesta"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetricasRespuesta": {
        "properties": {
          "cuentas": {
            "type": "integer",
            "title": "Cuentas"
          },
          "capital_total": {
            "type": "string",
            "title": "Capital Total"
          },
          "cuota_minima": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuota Minima"
          },
          "cuota_maxima": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuota Maxima"
          },
          "cuentas_por_segmento": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cuentas Por Segmento"
          },
          "adicionales": {
            "additionalProperties": true,
            "type": "object",
            "title": "Adicionales"
          }
        },
        "type": "object",
        "required": [
          "cuentas",
          "capital_total",
          "cuota_minima",
          "cuota_maxima",
          "cuentas_por_segmento",
          "adicionales"
        ],
        "title": "MetricasRespuesta"
      },
...
      "PaginaRespuesta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limite": {
            "type": "integer",
            "title": "Limite"
          },
          "desplazamiento": {
            "type": "integer",
            "title": "Desplazamiento"
          },
          "suficiente": {
            "type": "boolean",
            "title": "Suficiente"
          },
          "productos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Productos"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limite",
          "desplazamiento",
          "suficiente",
          "productos"
        ],
        "title": "PaginaRespuesta"
      },
...
      "ResumenRespuesta": {
        "properties": {
          "disponibles": {
            "type": "integer",
            "title": "Disponibles"
          },
          "solicitados": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solicitados"
          },
          "suficiente": {
            "type": "boolean",
            "title": "Suficiente"
          },
          "universo": {
            "$ref": "#/components/schemas/BloqueRespuesta"
          },
          "seleccion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BloqueRespuesta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "disponibles",
          "solicitados",
          "suficiente",
          "universo",
          "seleccion"
        ],
        "title": "ResumenRespuesta"
      },
...
      "SegmentacionRespuesta": {
        "properties": {
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "grupos": {
            "items": {
              "$ref": "#/components/schemas/GrupoRespuesta"
            },
            "type": "array",
            "title": "Grupos"
          }
        },
        "type": "object",
        "required": [
          "campo",
          "grupos"
        ],
        "title": "SegmentacionRespuesta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}